upload profile pic from complpete profile, and append postedActivity to div
