+++ backend/src/views/profile/jobCompleteProfile.php
@@ -116,6 +116,8 @@
                           <!--First tab-->
                           <div id="tab_profile" class="tab-pane">
 
+                            <?php require_once 'views/forms/profilePicture.php'; ?>
+
                             <div class="form-horizontal">
                               
                               <?php require_once 'views/forms/basicInformation.php'; ?>
