+++ backend/src/views/setting/setting.php
@@ -86,31 +86,7 @@
                           </h3>
                         </div> <!-- /.heading-block -->
 
-                        <form action="profile/updatePic" class="form-horizontal" method="POST">
-
-                          <div class="form-group">
-
-                            <label class="col-md-3 control-label">Avatar</label>
-
-                            <div class="col-md-7">
-
-                              <div class="fileinput fileinput-new" data-provides="fileinput">
-                                <div class="fileinput-preview thumbnail" data-trigger="fileinput" style="max-width: 200px;">
-                                  
-                                  <img id="target" alt="Avatar" src="uploads/profilePictures/<?= $_SESSION['username'] ?>.jpg" >
-                                </div>
-                                <div>
-
-                                <input id="src" type="file"/> 
-                                  
-                                </div>
-                              </div> <!-- /.fileupload -->
-
-                            </div> <!-- /.col -->
-
-                          </div> <!-- /.form-group -->
-
-                        </form>
+                        <?php require_once 'views/forms/profilePicture.php'; ?>
 
                         <div class="heading-block">
                           <h3>
@@ -375,68 +351,6 @@
     <?php include_once "static/js/updateUserProfileSettingFunctions.php"; ?>
 
     
-<script type="text/javascript">
-  function uploadProfilePic(){
-            //var _progress = document.getElementById('_progress'); 
-            if(src.files.length === 0){
-              
-              return false ;
-            }
-            else if (src.files['0'].size > 2015000) {
-             
-                  $.niftyNoty({ 
-                    type:"danger",
-                    icon:"fa fa-check fa-lg",
-                    title:"File Upload Error",
-                    message:"File size is too large",
-                    focus: true,
-                    container:"floating",
-                    timer:4000
-                  });
-              
-              return false ;
-            } 
-            else {
-              var data = new FormData();
-              data.append('file', src.files[0]);
-              var request = new XMLHttpRequest();
-              var responceTx = "";
-              request.onreadystatechange = function(){
-                if(request.readyState == 4){
-                  responceTx = request.response;
-                  $.niftyNoty({ 
-                    type:"danger",
-                    icon:"fa fa-check fa-lg",
-                    title:"Profile Pic Change Successfully",
-                    message:responceTx,
-                    focus: true,
-                    container:"floating",
-                    timer:4000
-                  });
-                }
-              };
-            }
-            request.upload.addEventListener('progress', function(e){
-              //_progress.style.width = Math.ceil(e.loaded/e.total) * 100 + '%';
-            }, false);  
-            request.open('POST', '<?= $baseUrl ?>fileUpload/profilePic');
-            request.send(data);
-          }
 
-  function showImage() {
-    var src = document.getElementById("src");
-    var target = document.getElementById("target");
-  var fr=new FileReader();
-  // when image is loaded, set the src of the image where you want to display it
-  fr.onload = function(e) { target.src = this.result; };
-  src.addEventListener("change",function() {
-    // fill fr with image data    
-    fr.readAsDataURL(src.files[0]);
-    uploadProfilePic();
-  });
-}
-showImage();
-
-</script>  
   </body>
 </html>
